+++ tests/acceptance/brocfile-smoke-test-slow.js
@@ -1,0 +1,95 @@
+'use strict';
+
+var tmp        = require('../helpers/tmp');
+var conf       = require('../helpers/conf');
+var Promise    = require('../../lib/ext/promise');
+var path       = require('path');
+var rimraf     = Promise.denodeify(require('rimraf'));
+var fs         = require('fs');
+var ncp        = Promise.denodeify(require('ncp'));
+var runCommand = require('../helpers/run-command');
+
+var appName  = 'some-cool-app';
+
+var rootPath = process.cwd();
+
+function buildApp(appName) {
+  return runCommand(path.join('..', 'bin', 'ember'), 'new', appName, {
+    onOutput: function() {
+      return; // no output for initial application build
+    }
+  });
+}
+
+function copyFixtureFiles(sourceDir) {
+  return ncp(path.join(rootPath, 'tests', 'fixtures', 'brocfile-tests', sourceDir), '.', {
+    clobber: true,
+    stopOnErr: true
+  });
+}
+
+describe('Acceptance: brocfile-smoke-test', function() {
+  before(function() {
+    this.timeout(360000);
+
+    tmp.setup('./common-tmp');
+    process.chdir('./common-tmp');
+
+    conf.setup();
+    return buildApp(appName)
+      .then(function() {
+        return rimraf(path.join(appName, 'node_modules', 'ember-cli'));
+      });
+  });
+
+  after(function() {
+    tmp.teardown('./common-tmp');
+    conf.restore();
+  });
+
+  beforeEach(function() {
+    this.timeout(10000);
+    tmp.setup('./tmp');
+    return ncp('./common-tmp/' + appName, './tmp/' + appName, {
+      clobber: true,
+      stopOnErr: true
+    })
+    .then(function() {
+      process.chdir('./tmp');
+
+      var appsECLIPath = path.join(appName, 'node_modules', 'ember-cli');
+      var pwd = process.cwd();
+
+      fs.symlinkSync(path.join(pwd, '..'), appsECLIPath);
+
+      process.chdir(appName);
+    });
+  });
+
+  afterEach(function() {
+    tmp.teardown('./tmp');
+  });
+
+  it('using wrapInEval: true', function() {
+    console.log('    running the slow end-to-end it will take some time');
+
+    this.timeout(450000);
+
+    return copyFixtureFiles('wrap-in-eval')
+      .then(function() {
+        return runCommand(path.join('.', 'node_modules', 'ember-cli', 'bin', 'ember'), 'test').then(console.log);
+      });
+  });
+
+  it('default development build tests', function() {
+    console.log('    running the slow end-to-end it will take some time');
+
+    this.timeout(450000);
+
+    return copyFixtureFiles('default-development')
+    .then(function() {
+      return runCommand(path.join('.', 'node_modules', 'ember-cli', 'bin', 'ember'), 'test');
+    });
+  });
+});
+