+++ tests/acceptance/smoke-test-slow.js
@@ -3,7 +3,6 @@
 var tmp      = require('../helpers/tmp');
 var conf     = require('../helpers/conf');
 var Promise  = require('../../lib/ext/promise');
-var exec     = Promise.denodeify(require('child_process').exec);
 var path     = require('path');
 var rimraf   = Promise.denodeify(require('rimraf'));
 var fs       = require('fs');
@@ -11,18 +10,66 @@ var crypto   = require('crypto');
 var assert   = require('assert');
 var walkSync = require('walk-sync');
 var appName  = 'some-cool-app';
+var ncp      = Promise.denodeify(require('ncp'));
+var runCommand = require('../helpers/run-command');
+
+function assertTmpEmpty() {
+  var paths = walkSync('tmp')
+    .filter(function(path) {
+      return !path.match(/output\//);
+    });
+
+  assert(paths.length === 0, 'tmp/ should be empty after `ember` tasks. Contained: ' + paths.join('\n'));
+}
+
+function buildApp(appName) {
+  return runCommand(path.join('..', 'bin', 'ember'), 'new', appName, {
+    onOutput: function() {
+      return; // no output for initial application build
+    }
+  });
+}
 
 describe('Acceptance: smoke-test', function() {
-  before(conf.setup);
+  before(function() {
+    this.timeout(360000);
 
-  after(conf.restore);
+    tmp.setup('./common-tmp');
+    process.chdir('./common-tmp');
+
+    conf.setup();
+    return buildApp(appName)
+      .then(function() {
+        return rimraf(path.join(appName, 'node_modules', 'ember-cli'));
+      });
+  });
+
+  after(function() {
+    tmp.teardown('./common-tmp');
+    conf.restore();
+  });
 
   beforeEach(function() {
+    this.timeout(10000);
     tmp.setup('./tmp');
-    process.chdir('./tmp');
+    return ncp('./common-tmp/' + appName, './tmp/' + appName, {
+      clobber: true,
+      stopOnErr: true
+    })
+    .then(function() {
+      process.chdir('./tmp');
+
+      var appsECLIPath = path.join(appName, 'node_modules', 'ember-cli');
+      var pwd = process.cwd();
+
+      fs.symlinkSync(path.join(pwd, '..'), appsECLIPath);
+
+      process.chdir(appName);
+    });
   });
 
   afterEach(function() {
+    assertTmpEmpty();
     tmp.teardown('./tmp');
   });
 
@@ -31,21 +78,7 @@ describe('Acceptance: smoke-test', function() {
 
     this.timeout(450000);
 
-    var appsECLIPath = path.join(appName, 'node_modules', 'ember-cli');
-
-    return exec('pwd').then(function(pwd) {
-      return exec(path.join('..', 'bin', 'ember') + ' new ' + appName).then(function() {
-        return rimraf(appsECLIPath).then(function() {
-          fs.symlinkSync(path.join(pwd, '..'), appsECLIPath);
-
-          process.chdir(appName);
-
-          return exec(path.join('.', 'node_modules', 'ember-cli', 'bin', 'ember') + ' test').then(console.log);
-        });
-      });
-    }).finally(function() {
-      console.log('done!');
-    });
+    return runCommand(path.join('.', 'node_modules', 'ember-cli', 'bin', 'ember'), 'test').then(console.log);
   });
 
   it('ember new foo, build production and verify fingerprint', function() {
@@ -53,51 +86,35 @@ describe('Acceptance: smoke-test', function() {
 
     this.timeout(360000);
 
-    var appsECLIPath = path.join(appName, 'node_modules', 'ember-cli');
-    var globalPwd = null;
+    return runCommand(path.join('.', 'node_modules', 'ember-cli', 'bin', 'ember'), 'build', '--environment=production')
+      .then(function() {
+        var dirPath = path.join('.', 'dist', 'assets');
+        var dir = fs.readdirSync(dirPath);
+        var files = [];
 
-    return exec('pwd').then(function(pwd) {
-      globalPwd = pwd;
+        dir.forEach(function (filepath) {
+          if (filepath === '.gitkeep') {
+            return;
+          }
 
-      return exec(path.join('..', 'bin', 'ember') + ' new ' + appName);
-    }).then(function() {
-      return rimraf(appsECLIPath);
-    }).then(function() {
-      fs.symlinkSync(path.join(globalPwd, '..'), appsECLIPath);
+          files.push(filepath);
 
-      process.chdir(appName);
+          var file = fs.readFileSync(path.join(dirPath, filepath), { encoding: 'utf8' });
 
-      return exec(path.join('.', 'node_modules', 'ember-cli', 'bin', 'ember') + ' build --environment=production');
-    }).then(function() {
-      var dirPath = path.join('.', 'dist', 'assets');
-      var dir = fs.readdirSync(dirPath);
-      var files = [];
+          var md5 = crypto.createHash('md5');
+          md5.update(file);
+          var hex = md5.digest('hex');
 
-      dir.forEach(function (filepath) {
-        if (filepath === '.gitkeep') {
-          return;
-        }
+          var possibleNames = [appName + '-' + hex + '.js', appName + '-' + hex + '.css', 'vendor-' + hex + '.js', 'vendor-' + hex + '.css'];
+          assert(possibleNames.indexOf(filepath) > -1);
+        });
 
-        files.push(filepath);
+        var indexHtml = fs.readFileSync(path.join('.', 'dist', 'index.html'), { encoding: 'utf8' });
 
-        var file = fs.readFileSync(path.join(dirPath, filepath), { encoding: 'utf8' });
-
-        var md5 = crypto.createHash('md5');
-        md5.update(file);
-        var hex = md5.digest('hex');
-
-        var possibleNames = [appName + '-' + hex + '.js', appName + '-' + hex + '.css', 'vendor-' + hex + '.js', 'vendor-' + hex + '.css'];
-        assert(possibleNames.indexOf(filepath) > -1);
+        files.forEach(function (filename) {
+          assert(indexHtml.indexOf(filename) > -1);
+        });
       });
-
-      var indexHtml = fs.readFileSync(path.join('.', 'dist', 'index.html'), { encoding: 'utf8' });
-
-      files.forEach(function (filename) {
-        assert(indexHtml.indexOf(filename) > -1);
-      });
-    }).finally(function() {
-      console.log('done');
-    });
   });
 
   it('ember new foo, build development, and verify generated files', function() {
@@ -105,28 +122,29 @@ describe('Acceptance: smoke-test', function() {
 
     this.timeout(360000);
 
-    var appsECLIPath = path.join(appName, 'node_modules', 'ember-cli');
-    var globalPwd = null;
+    return runCommand(path.join('.', 'node_modules', 'ember-cli', 'bin', 'ember'), 'build')
+      .then(function() {
+        var dirPath = path.join('.', 'dist');
+        var paths = walkSync(dirPath);
 
-    return exec('pwd').then(function(pwd) {
-      globalPwd = pwd;
+        assert(paths.length < 20, 'expected fewer than 20 files in dist, found ' + paths.length);
+      });
+  });
 
-      return exec(path.join('..', 'bin', 'ember') + ' new ' + appName);
-    }).then(function() {
-      return rimraf(appsECLIPath);
-    }).then(function() {
-      fs.symlinkSync(path.join(globalPwd, '..'), appsECLIPath);
+  it('ember new foo, server, SIGINT clears tmp/', function() {
+    console.log('    running the slow build tests');
 
-      process.chdir(appName);
+    this.timeout(360000);
 
-      return exec(path.join('.', 'node_modules', 'ember-cli', 'bin', 'ember') + ' build');
-    }).then(function() {
-      var dirPath = path.join('.', 'dist');
-      var paths = walkSync(dirPath);
-
-      assert(paths.length < 20, 'expected fewer than 20 files in dist, found ' + paths.length);
-    }).finally(function() {
-      console.log('done');
-    });
+    return runCommand(path.join('.', 'node_modules', 'ember-cli', 'bin', 'ember'), 'server', {
+        onOutput: function(string, process) {
+          if (string.match(/Build successful/)) {
+            process.kill('SIGINT');
+          }
+        }
+      })
+      .catch(function() {
+        // just eat the rejection as we are testing what happens
+      });
   });
 });