+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### 0.0.34
 
-* [BUGFIX] broccoli-es6-safe-recast now once again has one-at-a-time semantics for incremental rebuilds
+* [BUGFIX] broccoli-es6-safe-recast now once again has one-at-a-time semantics this improves incremental rebuild performance
 * [BUGFIX] upgrade broccoli-sane-watcher to include better error messages when attempting to watch non-existent files
 * [ENHANCEMENT] Allow opting out of `ES3SafeFilter`. [#966](https://github.com/stefanpenner/ember-cli/pull/966)
 * [ENHANCEMENT] Provide `--watcher` option for switching between polling and events-based file watching. [#970](https://github.com/stefanpenner/ember-cli/pull/970)