+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # ember-cli Changelog
 
+* [BUGFIX] Ensure that vendored JS files are concatted in a safe way (to prevent issues with ASI). [#988](https://github.com/stefanpenner/ember-cli/pull/988)
+
 ### 0.0.34
 
 * [BUGFIX] broccoli-es6-safe-recast now once again has one-at-a-time semantics this improves incremental rebuild performance